rename component files
