+++ ui/components/app/confirm-page-container/flask/snap-insight.js
